test: add DefinitionValidatorTestDefinition class for improved validation testing
